Dev: Remove hooks
