+++ pages/user/balance/add.js
@@ -11,6 +11,31 @@ export default function Index(props) {
   const [preferenceId, setPreferenceId] = useState(null)
   const initialized = useRef(false)
 
+
+   function useScript(url : string) {
+    const [loaded, setLoaded] = useState(false)
+    useEffect(() => {
+      const existingScript = document.querySelector(`[src="${url}"]`)
+  
+      if (existingScript) {
+        setLoaded(true)
+      } else {
+        const script = document.createElement("script")
+        script.src = url
+        script.async = true
+        script.onload = () => {
+          setLoaded(true)
+        }
+  
+        document.body.appendChild(script)
+      }
+    }, [url])
+  
+    return {
+      loaded
+    }
+  }
+
   const { loaded } = useScript("https://sdk.mercadopago.com/js/v2")
 
   useEffect(() => {
